requirements.txt
requests>=2.31.0
pyyaml>=6.0
jinja2>=3.1.2
anthropic==0.40.0
pytest>=7.0.0
